requirements.txt
# core 
langchain==0.3.13
langchain_openai==0.2.14
openai==1.58.1
pydantic==2.10.4

## formatting
black==24.10.0

## example dependencies
numpy==2.2.1
wikipedia==1.4.0
duckduckgo-search==7.1.1
openmeteo_requests==1.3.0
requests_cache==1.2.1
retry_requests==2.0.0

## UI dependencies
streamlit==1.41.1
